move _perform_request() back into requests.py
